+++ tsup.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
     'src/cli.ts',
     'src/runtime/index.ts',
     'src/runtime/advanced.ts',
+    'src/runtime/kafkajs.ts',
     'src/runtime/platformatic.ts'
   ],
   external: ['@platformatic/kafka', 'avsc', 'kafkajs', 'zod'],